+++ sidertests/__init__.py
@@ -1,4 +1,5 @@
 import doctest
+import os
 from attest import Tests
 from . import session, types, hash, list, set
 
@@ -21,4 +22,11 @@ def doctest_types():
 def doctest_datetime():
     from sider import datetime
     assert 0 == doctest.testmod(datetime)[0]
+
+
+@tests.test
+def print_version():
+    from sider.version import VERSION
+    printed_version = os.popen('python -m sider.version').read().strip()
+    assert printed_version == VERSION